consolidate server tests
